backend/requirements.txt
flask==2.1.0
flask-cors==3.0.10
torch==2.1.0
diffusers==0.23.0
pillow==9.5.0
transformers==4.35.0